+++ .idea/dictionaries/Dell.xml
@@ -3,6 +3,7 @@
     <words>
       <w>tasker</w>
       <w>tmdb</w>
+      <w>tmdblib</w>
     </words>
   </dictionary>
 </component>
