--- app/src/main/java/lampa/test/tmdblib/Koin/viewModel.kt
@@ -1,9 +1,0 @@
-package lampa.test.tmdblib.Koin
-
-import lampa.test.tmdblib.model.viewmodel.VideoViewModel
-import org.koin.androidx.viewmodel.dsl.viewModel
-import org.koin.dsl.module
-
-val viewModel = module {
-    viewModel { VideoViewModel() }
-}
