+++ app/src/main/java/lampa/test/tmdblib/model/viewmodel/repository/data/MovieTmdbData.kt
@@ -1,0 +1,8 @@
+package lampa.test.tmdblib.model.viewmodel.repository.data
+
+    data class MovieTmdbData (
+
+        val page: Int,
+        val total_pages: Int,
+        val results: List<MovieResultsTmdbData>
+    )
