+++ app/src/main/java/lampa/test/tmdblib/model/viewmodel/repository/data/PostMovieRatingData.kt
@@ -1,0 +1,6 @@
+package lampa.test.tmdblib.model.viewmodel.repository.data
+
+data class PostMovieRatingData (
+
+    val value:Float
+)
