+++ app/src/main/java/lampa/test/tmdblib/model/viewmodel/repository/data/PostResponseData.kt
@@ -1,0 +1,7 @@
+package lampa.test.tmdblib.model.viewmodel.repository.data
+
+data class PostResponseData (
+
+    val status_code: Int,
+    val status_message: String
+)
