+++ app/src/main/java/lampa/test/tmdblib/model/viewmodel/repository/data/SessionTmdbData.kt
@@ -1,0 +1,8 @@
+package lampa.test.tmdblib.model.viewmodel.repository.data
+
+data class SessionTmdbData(
+
+    val success: Boolean,
+    val guest_session_id: String,
+    val expires_at: String
+)
