+++ app/src/main/java/lampa/test/tmdblib/model/viewmodel/repository/data/WrapperMovieData.kt
@@ -1,0 +1,8 @@
+package lampa.test.tmdblib.model.viewmodel.repository.data
+
+    data class WrapperMovieData (
+
+        var showAllOrAddToShow:Int,
+        val movieTmdbData: MovieTmdbData,
+        val toLikeList: Boolean
+    )
