+++ app/src/main/java/lampa/test/tmdblib/model/viewmodel/repository/internet/parser/callBack/CallBackPageFromParser.kt
@@ -1,0 +1,7 @@
+package lampa.test.tmdblib.model.viewmodel.repository.internet.parser.callBack
+
+interface CallBackPageFromParser {
+
+    fun onPageFind(link: String)
+    fun onPageNotFound()
+}
