+++ app/src/main/java/lampa/test/tmdblib/model/viewmodel/repository/internet/parser/callBack/CallBackVideoFromParser.kt
@@ -1,0 +1,6 @@
+package lampa.test.tmdblib.model.viewmodel.repository.internet.parser.callBack
+
+interface CallBackVideoFromParser {
+
+    fun onVideoFind(link: String)
+}
