--- app/src/main/java/lampa/test/tmdblib/repository/data/MovieTmdbData.kt
@@ -1,8 +1,0 @@
-package lampa.test.tmdblib.repository.data
-
-    data class MovieTmdbData (
-
-        val page: Int,
-        val total_pages: Int,
-        val results: List<MovieResultsTmdbData>
-    )
