--- app/src/main/java/lampa/test/tmdblib/repository/data/PostMovieRatingData.kt
@@ -1,6 +1,0 @@
-package lampa.test.tmdblib.repository.data
-
-data class PostMovieRatingData (
-
-    val value:Float
-)
