--- app/src/main/java/lampa/test/tmdblib/repository/data/PostResponseData.kt
@@ -1,7 +1,0 @@
-package lampa.test.tmdblib.repository.data
-
-data class PostResponseData (
-
-    val status_code: Int,
-    val status_message: String
-)
