--- app/src/main/java/lampa/test/tmdblib/repository/data/WrapperMovieData.kt
@@ -1,8 +1,0 @@
-package lampa.test.tmdblib.repository.data
-
-    data class WrapperMovieData (
-
-        var showAllOrAddToShow:Int,
-        val movieTmdbData: MovieTmdbData,
-        val toLikeList: Boolean
-    )
