--- app/src/main/java/lampa/test/tmdblib/repository/internet/parser/callBack/CallBackPageFromParser.kt
@@ -1,6 +1,0 @@
-package lampa.test.tmdblib.repository.internet.parser.callBack
-
-interface CallBackPageFromParser {
-
-    fun onPageFind(link: String)
-}
