--- app/src/main/java/lampa/test/tmdblib/repository/internet/parser/callBack/CallBackVideoFromParser.kt
@@ -1,6 +1,0 @@
-package lampa.test.tmdblib.repository.internet.parser.callBack
-
-interface CallBackVideoFromParser {
-
-    fun onVideoFind(link: String)
-}
