+++ app/src/main/java/lampa/test/tmdblib/view/activity/VideoActivity.kt
@@ -7,13 +7,15 @@ import android.webkit.JavascriptInterface
 import android.webkit.WebView
 import android.webkit.WebViewClient
 import android.widget.MediaController
+import kotlinx.android.synthetic.*
 import kotlinx.android.synthetic.main.activity_video.*
 import lampa.test.tmdblib.R
-import lampa.test.tmdblib.repository.data.MovieResultsTmdbData
-import lampa.test.tmdblib.repository.internet.parser.JavaScriptParserPage
-import lampa.test.tmdblib.repository.internet.parser.JavaScriptParserVideo
-import lampa.test.tmdblib.repository.internet.parser.callBack.CallBackPageFromParser
-import lampa.test.tmdblib.repository.internet.parser.callBack.CallBackVideoFromParser
+import lampa.test.tmdblib.model.viewmodel.repository.data.MovieResultsTmdbData
+import lampa.test.tmdblib.model.viewmodel.repository.internet.parser.JavaScriptParserPage
+import lampa.test.tmdblib.model.viewmodel.repository.internet.parser.JavaScriptParserVideo
+import lampa.test.tmdblib.model.viewmodel.repository.internet.parser.callBack.CallBackPageFromParser
+import lampa.test.tmdblib.model.viewmodel.repository.internet.parser.callBack.CallBackVideoFromParser
+import lampa.test.tmdblib.util.toast.toast
 import lampa.test.tmdblib.view.activity.base.BaseActivity
 
 
@@ -22,30 +24,44 @@ class VideoActivity : BaseActivity(), CallBackVideoFromParser, CallBackPageFromP
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_video)
-
-        (intent.extras?.get("content") as MovieResultsTmdbData).let {
-            findPageUtlWithVideoView(it)
-        }
+        findPageUrlWithVideoView((intent.extras?.get("content") as MovieResultsTmdbData))
     }
 
     @JavascriptInterface
-    fun findPageUtlWithVideoView(content: MovieResultsTmdbData) {
+    fun findPageUrlWithVideoView(content: MovieResultsTmdbData) {
 
-        webView.settings.javaScriptEnabled = true
-        webView.addJavascriptInterface(JavaScriptParserPage(this), "PAGE")
-        webView.setWebViewClient(object : WebViewClient() {
-            override fun onPageFinished(view: WebView, url: String) {
-                webView.loadUrl("javascript:window.PAGE.processHTML('<head>'+document.getElementsByTagName('html')[0].innerHTML+'</head>');")
-            }
-        })
-        webView.getSettings().setUserAgentString("Chrome/41.0.2228.0 Safari/537.36")
+        runOnUiThread {
+            webView.settings.javaScriptEnabled = true
+            webView.addJavascriptInterface(JavaScriptParserPage(this), "PAGE")
+            webView.setWebViewClient(object : WebViewClient() {
+                override fun onPageFinished(view: WebView, url: String) {
+                    webView.loadUrl("javascript:window.PAGE.processHTML('<head>'+document.getElementsByTagName('html')[0].innerHTML+'</head>');")
+                }
+            })
+            webView.getSettings().setUserAgentString("Chrome/41.0.2228.0 Safari/537.36")
 
-        Log.v("112233", "load " + "https://filmix.co/search/" + content.title + "-" + content.release_date.substring(0,4))
-        webView.loadUrl("https://filmix.co/search/" + content.title + "-" + content.release_date.substring(0,4))
-    }//+ content.title + "-" + content.release_date
+            Log.v(
+                "112233",
+                "load " + "https://filmix.co/search/" + content.title + "-" + content.release_date.substring(
+                    0,
+                    4
+                )
+            )
+            webView.loadUrl(
+                "https://filmix.co/search/" + content.title + "-" + content.release_date.substring(
+                    0,
+                    4
+                )
+            )
+        }
+    }
 
     override fun onPageFind(link: String) {
         findVideoUrlWithVideoView(link)
+    }
+
+    override fun onPageNotFound() {
+        toast("ошибка 101, попробуйте перезапустить приложение")
     }
 
     @JavascriptInterface
@@ -72,6 +88,12 @@ class VideoActivity : BaseActivity(), CallBackVideoFromParser, CallBackPageFromP
             val mediaController = MediaController(this)
             videoView.setMediaController(mediaController)
             mediaController.setAnchorView(videoView)
+            videoView.start()
         }
+    }
+
+    override fun onDestroy() {
+        clearFindViewByIdCache()
+        super.onDestroy()
     }
 }
