+++ app/src/main/java/lampa/test/tmdblib/view/fragments/callback/CallBackFromMainFToActivity.kt
@@ -1,6 +1,6 @@
 package lampa.test.tmdblib.view.fragments.callback
 
-import lampa.test.tmdblib.repository.data.MovieResultsTmdbData
+import lampa.test.tmdblib.model.viewmodel.repository.data.MovieResultsTmdbData
 
 interface CallBackFromMainFToActivity {
